+++ Common/SpriteRenderInfo.h
@@ -7,7 +7,8 @@ class SpriteRenderInfo : public IBinarySerializable
 {
 public:
 	SpriteRenderInfo() noexcept;
-	SpriteRenderInfo(std::shared_ptr<Sprite> sprite, Box2i boundingBox, int sizeInUnits, int layerIndex) NOEXCEPT_WHEN_NDEBUG;
+	SpriteRenderInfo(std::shared_ptr<Sprite> sprite, 
+		Box2i boundingBox, float sizeInUnits, int layerIndex) NOEXCEPT_WHEN_NDEBUG;
 	SpriteRenderInfo(const SpriteRenderInfo& spriteRenderInfo) noexcept;
 
 	SpriteRenderInfo& operator=(const SpriteRenderInfo& right) noexcept;
@@ -20,7 +21,7 @@ public:
 	HBITMAP GetBitmap() const noexcept;
 	Box2i GetBoundingBox() const noexcept;
 	Vector2i GetPosition() const noexcept;
-	int GetSizeInUnits() const noexcept;
+	float GetSizeInUnits() const noexcept;
 	int GetLayerIndex() const noexcept;
 
 	void SerializeToOpenedFile(std::ofstream& file) const override;
@@ -30,6 +31,6 @@ private:
 	std::shared_ptr<Sprite> m_sprite;
 	Box2i m_boundingBox;
 	Vector2i m_positionInUnits;
-	int m_sizeInUnits{ 0 };
+	float m_sizeInUnits{ 0 };
 	int m_layerIndex{ 0 };
 };
